inline `sequence64` source & unit test files
