--- packages/orbital-sequences/src/sequence64/increment/increment.ts
@@ -1,1 +1,0 @@
-export const increment = 1n
